Add 4.21 solutions

+++ fullstack-open-exercises/part4/controllers/blogs.js
@@ -29,7 +29,6 @@ blogsRouter.post("/", async (request, response, next) => {
     return response.status(401).json({ error: "token missing or invalid" });
   }
   const user = await User.findById(decodedToken.id);
-
   const blog = new Blog({
     title: body.title,
     author: body.author,
@@ -46,8 +45,19 @@ blogsRouter.post("/", async (request, response, next) => {
 });
 
 blogsRouter.delete("/:id", async (request, response, next) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  // check if token id matches the user id?
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+  const blog = await Blog.findById(request.params.id);
+  if (blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    return response.status(401).json({ error: "incorrect user" });
+  }
 });
 
 blogsRouter.put("/:id", async (request, response, next) => {
